+++ rect.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Rect struct {
-	X1, Y1, X2, Y2 float32
+	X1, Y1, X2, Y2 float64
 	clean          bool
 }
 
-func NewRect(x1, y1, x2, y2 float32) Rect {
+func NewRect(x1, y1, x2, y2 float64) Rect {
 	var r Rect
 
 	r.X1 = x1
@@ -45,7 +45,7 @@ func (rect *Rect) SetRect(r Rect) {
 	rect.clean = false
 }
 
-func (rect *Rect) Set(x1, y1, x2, y2 float32) {
+func (rect *Rect) Set(x1, y1, x2, y2 float64) {
 	rect.X1 = x1
 	rect.X2 = x2
 	rect.Y1 = y1
@@ -53,7 +53,7 @@ func (rect *Rect) Set(x1, y1, x2, y2 float32) {
 	rect.clean = false
 }
 
-func (rect *Rect) SetRadius(x, y, r float32) {
+func (rect *Rect) SetRadius(x, y, r float64) {
 	rect.X1 = x - r
 	rect.X2 = x + r
 	rect.Y1 = y - r
@@ -61,7 +61,7 @@ func (rect *Rect) SetRadius(x, y, r float32) {
 	rect.clean = false
 }
 
-func (rect *Rect) Encapsulate(x, y float32) {
+func (rect *Rect) Encapsulate(x, y float64) {
 	if rect.clean {
 		rect.X1 = x
 		rect.X2 = x
@@ -84,7 +84,7 @@ func (rect *Rect) Encapsulate(x, y float32) {
 	}
 }
 
-func (rect *Rect) TestPoint(x, y float32) bool {
+func (rect *Rect) TestPoint(x, y float64) bool {
 	if x >= rect.X1 && x < rect.X2 && y >= rect.Y1 && y < rect.Y2 {
 		return true
 	}
@@ -93,8 +93,8 @@ func (rect *Rect) TestPoint(x, y float32) bool {
 }
 
 func (rect *Rect) Intersect(r *Rect) bool {
-	if math.Abs(float64(rect.X1+rect.X2-r.X1-r.X2)) < float64(rect.X2-rect.X1+r.X2-r.X1) {
-		if math.Abs(float64(rect.Y1+rect.Y2-r.Y1-r.Y2)) < float64(rect.Y2-rect.Y1+r.Y2-r.Y1) {
+	if math.Abs(rect.X1+rect.X2-r.X1-r.X2) < rect.X2-rect.X1+r.X2-r.X1 {
+		if math.Abs(rect.Y1+rect.Y2-r.Y1-r.Y2) < rect.Y2-rect.Y1+r.Y2-r.Y1 {
 			return true
 		}
 	}
